made a few folders refactories
